Use algorithm structure with generators

+++ src/options/options.rs
@@ -14,7 +14,7 @@ pub struct Options {
     #[structopt(short, long, default_value = "1", help = "Number of passwords to generate")]
     pub num: u8,
     #[structopt(short, long, default_value = "3", help = "Number of password tokens")]
-    pub tokens: u8
+    pub tokens: usize
 }
 
 impl Options {
